Use ga4gh-common for setup and utils
