+++ .flake8
@@ -1,0 +1,2 @@
+[flake8]
+exclude = ga4gh/*_pb2.py,ga4gh/datamodel/obo_parser.py,docs,ez_setup.py
