+++ dev-requirements.txt
@@ -16,16 +16,18 @@ PyYAML==3.11
 # do not import them in any code that can be reached via 
 # an executable that we ship
 
+mock
+nose
+pep8
+flake8
+coverage
+
 PyVCF==0.6.7
-coverage==3.7.1
-flake8==2.3.0  # Set due to conflict with pep8 in version 2.4.0
 freezegun==0.3.6
 guppy==0.1.10
-mock==1.0.0
-nose==1.3.7
-pep8==1.6.2
 snakefood==1.4
 Sphinx==1.3.1
+
 # Requirements for OIDC provider
 pyaml==15.03.1
 cherrypy==3.2.4
