+++ ga4gh/cli/__init__.py
@@ -5,9 +5,6 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import argparse
-import operator
-
 import ga4gh
 import ga4gh.protocol as protocol
 
@@ -18,52 +15,6 @@ import ga4gh.protocol as protocol
 # This should be removed once pysam input sanitisation has been
 # implemented.
 AVRO_LONG_MAX = 2**31 - 1
-
-
-class SortedHelpFormatter(argparse.HelpFormatter):
-    """
-    An argparse HelpFormatter that sorts the flags and subcommands
-    in alphabetical order
-    """
-    def add_arguments(self, actions):
-        """
-        Sort the flags alphabetically
-        """
-        actions = sorted(
-            actions, key=operator.attrgetter('option_strings'))
-        super(SortedHelpFormatter, self).add_arguments(actions)
-
-    def _iter_indented_subactions(self, action):
-        """
-        Sort the subcommands alphabetically
-        """
-        try:
-            get_subactions = action._get_subactions
-        except AttributeError:
-            pass
-        else:
-            self._indent()
-            if isinstance(action, argparse._SubParsersAction):
-                for subaction in sorted(
-                        get_subactions(), key=lambda x: x.dest):
-                    yield subaction
-            else:
-                for subaction in get_subactions():
-                    yield subaction
-            self._dedent()
-
-
-def addSubparser(subparsers, subcommand, description):
-    parser = subparsers.add_parser(
-        subcommand, description=description, help=description)
-    return parser
-
-
-def createArgumentParser(description):
-    parser = argparse.ArgumentParser(
-        description=description,
-        formatter_class=SortedHelpFormatter)
-    return parser
 
 
 def addVersionArgument(parser):
