+++ oidc-provider/simple_op/src/provider/authn/__init__.py
@@ -3,6 +3,7 @@ from oic.utils.authn.user import UserAuthnMethod
 
 __author__ = 'regu0004'
 
+
 class AuthnModule(UserAuthnMethod):
 
     # override in subclass specifying suitable url endpoint to POST user input
