+++ requirements.txt
@@ -8,6 +8,8 @@
 # these libraries are the set listed by pipdeptree -f -w
 # that are dependencies of libraries listed in the next section
 
+ga4gh-common==0.0.4
+
 Werkzeug==0.11.5
 MarkupSafe==0.23
 itsdangerous==0.24
