+++ scripts/bam2sam.py
@@ -5,7 +5,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-import utils
+import ga4gh_common.utils as utils
 
 
 @utils.Timed()
