+++ tests/unit/test_client.py
@@ -14,7 +14,7 @@ import ga4gh.protocol as protocol
 import ga4gh.backend as backend
 import ga4gh.client as client
 import ga4gh.datarepo as datarepo
-import tests.utils as utils
+import ga4gh_common.utils as utils
 import ga4gh.exceptions as exceptions
